requirements.txt
openai==1.76.0
python-dotenv==1.0.1
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-wtf==1.0.0
gunicorn==20.1.0
werkzeug==2.0.1
sqlalchemy==1.4.23